Removes obsolete simulation_input_output class
